add TMUX_PANES checks before splits

+++ home/user/bin/evm-tmux.sh
@@ -69,15 +69,15 @@ startMainnetTmux() {
 	tmux set-option -g mouse on		#deprecated: tmux set-option -g mouse-select-pane on
 
 	# Split the window into four panes: two panes on the left and two (one higher) on the right
-	if [ "$(tmux list-panes | wc -l)" -eq "1" ] ; then
+	if [ "$(tmux list-panes | wc -l)" -eq "1" ] && [ "${TMUX_PANE}" = "%1" ]; then
 		tmux select-pane -t 0
 		tmux split-window -h -t "${sTmuxSession}"
 	fi
-	if [ "$(tmux list-panes | wc -l)" -eq "2" ] ; then
+	if [ "$(tmux list-panes | wc -l)" -eq "2" ] && [ "${TMUX_PANE}" = "%2" ]; then
 		tmux select-pane -t 0 #-P 'p1'
 		tmux split-window -v -t "${sTmuxSession}" #-n 'p2'
 	fi
-	if [ "$(tmux list-panes | wc -l)" -eq "3" ] ; then
+	if [ "$(tmux list-panes | wc -l)" -eq "3" ] && [ "${TMUX_PANE}" = "%3" ]; then
 		tmux select-pane -t 2 #-n 'p3'
 		tmux split-window -v -l 3 #-p 90 #-t "${sTmuxSession}"
 	fi
@@ -88,7 +88,9 @@ startMainnetTmux() {
 		tmux send-keys -t "${sTmuxSession}:${sTmuxWindow}.2" "${sAlias3}" C-m #validator
 	fi
 	# alway echo this line at right bottom
-	tmux send-keys -t "${sTmuxSession}:${sTmuxWindow}.3" "echo -e 'to hide tmux (and keep running): press ctrl+b then d or enter \`tmux detach\`\nor to kill tmux: enter \`tmux kill-session -t evm\` to kill'" C-m
+	if [ "${TMUX_PANE}" = "%4" ]; then
+		tmux send-keys -t "${sTmuxSession}:${sTmuxWindow}.3" "echo -e 'to hide tmux (and keep running): press ctrl+b then d or enter \`tmux detach\`\nor to kill tmux: enter \`tmux kill-session -t evm\` to kill'" C-m
+	fi
 	tmux attach-session -t "${sTmuxSession}"	#tmux attach-session -t "${sTmuxSession}" -c "${sTmuxWindow}"	# Attach to the "${sTmuxSession}" session
 }
 upgradeBinTmuxEvmScript() {
